// src/llm-providers/openAI/enum/assistantAI.enum.ts
/* eslint-disable prettier/prettier */
export enum ModelType {
  GPT_3_5_TURBO = 'gpt-3.5-turbo',
  GPT_4 = 'gpt-4',
}

export enum QuestionType {
  TRUE_OR_FALSE = 'True or False',
  MULTIPLE_CHOICE = 'Multiple choice',
  OPEN_ENDED_LONG_FORM = 'Open ended long form',
  OPEN_ENDED_QUESTION = 'Open Ended Question',
  OPEN_ENDED_SHORT_FORM = 'Open ended short form',
  DX_QUEST = 'Dx Quest',
  PROBLEM_LIST_EVALUATION = 'Problem List Evaluation',
  MED_MATCH = 'Med-Match',
  BROAD_SCOPE_QUIZ = 'Broad Scope Quiz',
}

export enum ComponentType {
  ASSISTANT_AI = 'ASSISTANT_AI',
  EVENT_TEMPLATE = 'EVENT_TEMPLATE',
  OUTLINE_TO_JSON = 'OUTLINE_TO_JSON',
  SIMULATION_CONVERSION = 'SIMULATION_CONVERSION',
  CHAT_SIMULATION = 'CHAT_SIMULATION',
  QUIZ_AI = 'QUIZ_AI',
  EXAMPREP = 'EXAMPREP',
  OPENENDED = 'OPENENDED',
  CASE_RECALL = 'CASE_RECALL',
  MEDSYNOPSIS = 'MEDSYNOPSIS',
  USER_MEDSYNOPSIS = 'USER_MEDSYNOPSIS',
  DX_QUEST = 'DX QUEST',
  PROBLEM_LIST_EVALUATION = 'PROBLEM LIST EVALUATION',
  MED_MATCH = 'MED MATCH',
  BROAD_SCOPE_QUIZ = 'BROAD SCOPE QUIZ',
  CASE_PRESENTATION = 'CASE_PRESENTATION',
  SLIDE_PRESENTATION = 'SLIDE_PRESENTATION',
  PLAYGROUND_SLIDE_PRESENTATION = 'PLAYGROUND_SLIDE_PRESENTATION',
  CHAT_ASSISTANT = 'CHAT_ASSISTANT',
  BASIC_SCIENCE = 'BASIC SCIENCES',
  CLINICAL_SCIENCE = 'CLINICAL SCIENCES',
  GENERAL_TRIVIA = 'GENERAL TRIVIA',
  MEDICAL_TRIVIA = 'MEDICAL TRIVIA',
  USMLE_STEP1 = 'USMLE STEP 1',
  USMLE_STEP2 = 'USMLE STEP 2 CK',
  PLAB1 = 'PLAB 1',
  AMC1 = 'AMC 1',
  RACP1 = 'RACP 1',
  RACGP_AKT = 'RACGP (AKT)',
  RACGP_KFP = 'RACGP (KFP)',
  NCLEX_RN = 'NCLEX RN',
  NCLEX_PN = 'NCLEX PN',
  GROUP_QUESTIONS = 'GROUP QUESTIONS',
  CLINICAL_EXAM = 'CLINICAL EXAM',
  ENGAGEMENT = 'ENGAGEMENT',
  GENERATE_PODCAST = 'GENERATE_PODCAST',
  CLINICAL_EXAMS = 'CLINICAL_EXAMS',
}

export enum PromptType {
  AI_REWRITE = 'AI_REWRITE',
  AI_FEEDBACK = 'AI_FEEDBACK',
}

export enum FileExtType {
  PDF = '.pdf',
  DOCX = '.docx',
}

export enum GenderType {
  MALE = 'MALE',
  FEMALE = 'FEMALE',
}

export enum UserThreadType {
  CLINICAL_EXAM = 'clinicalExam',
  CASE_RECALL = 'caseRecall',
  CHAT_ASSISTANT = 'chatAssistant',
}